refactor: チェックイン詳細画面のロジックをHooksに分離
