+++ src/components/pages/CheckinDetail/CheckinDetail.tsx
@@ -1,53 +1,28 @@
-import React, { useState, useEffect, useCallback } from 'react'
+import React from 'react'
 import { View, Text, FlatList, ActivityIndicator } from 'react-native'
 import { Ionicons } from '@expo/vector-icons'
 import { Avatar, Icon } from 'react-native-elements'
-import { NavigationProp, ParamListBase, RouteProp } from '@react-navigation/native'
 import { ImageCarousel } from '@/components/organisms/ImageCarousel'
-import { useFoursquare } from '@/hooks/useFoursquare'
 import { COLORS } from '@/constants/Colors'
 import window from '@/constants/Layout'
-import { CheckinDetail } from '@/types/Foursquare'
 import { fontSize, fontColor, other } from '@/styles/styles'
 import CoinIcon from '@/components/CoinIcon.component'
-import { RootStackParamList } from '@/types'
-import useColorScheme from '@/hooks/useColorScheme'
 import MultipleName from '@/components/molecules/MultipleName'
 import CategoryIcon from '@/components/molecules/CategoryIcon'
 import Address from '@/components/organisms/Address'
 import { MayorIcon } from '@/components/organisms/MayorIcon'
 import { formatTimestamp, formatDistanceToNowForTimestamp, timestamp2Date } from '@/service/dateFns'
 import { generateImageUrl, removeShoutWith } from '@/service/utilFns'
-import { useNavigation, useRoute } from '@react-navigation/core'
+import { NoData } from '@/components/organisms/NoData'
+import useColorScheme from '@/hooks/useColorScheme'
+import { useCheckinDetail } from './useCheckinDetail'
 
 export const CheckinDetailScreen = () => {
-  const navigation = useNavigation<NavigationProp<ParamListBase>>()
-  const route = useRoute<RouteProp<RootStackParamList, 'CheckinDetail'>>()
   const colorScheme = useColorScheme()
-  const { item } = route.params
-  const [checkinDetail, setCheckinDetail] = useState<CheckinDetail & { images: string[] }>()
-  const { fetchCheckinDetails } = useFoursquare()
+  const { checkinDetail, loading } = useCheckinDetail()
+  if (loading) return <ActivityIndicator />
+  if (!checkinDetail) return <NoData />
 
-  const getCheckinDetails = useCallback(async () => {
-    const checkinDetail = await fetchCheckinDetails(item.id)
-    const images =
-      item.photos.count > 0
-        ? checkinDetail.photos.items.map((item) =>
-            generateImageUrl(item.prefix, item.suffix, 'cap400')
-          )
-        : ['']
-    setCheckinDetail({ ...checkinDetail, images })
-  }, [])
-
-  useEffect(() => {
-    if (!item) return
-    navigation.setOptions({ headerTitle: item.venue.name })
-    void getCheckinDetails()
-  }, [item])
-
-  if (!checkinDetail) {
-    return <ActivityIndicator />
-  }
   const itemRender = () => {
     return (
       <View style={{ paddingHorizontal: 8 }}>
@@ -64,7 +39,7 @@ export const CheckinDetailScreen = () => {
                   size={12}
                 />
               }
-              sum={checkinDetail?.likes.summary}
+              sum={checkinDetail.likes.summary}
             />
           )}
           {checkinDetail.with && (
@@ -115,7 +90,7 @@ export const CheckinDetailScreen = () => {
           </View>
         )}
         <View style={{ marginBottom: 8 }}>
-          <Text style={[fontColor.textSub]}>via: {item.source.name}</Text>
+          <Text style={[fontColor.textSub]}>via: {checkinDetail.source.name}</Text>
         </View>
 
         {checkinDetail.comments.items?.map((comment, i) => (
